+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [v1.1.3-dev] - 2022-07-15
+
+### Added
+
+- [changes in telecaller incentive report](https://gitlab.com/atri-tech/atri-maintainers/switch-my-loan/-/merge_requests/54)
+
+## [v1.1.2-dev] - 2022-07-15
+
+### Added
+
+- [sales team incentive report](https://gitlab.com/atri-tech/atri-maintainers/switch-my-loan/-/merge_requests/53)
+
 ## [v1.1.1-uat] - 2022-07-08
 ## [v1.1.1-dev] - 2022-07-08